# src/main/java/com/mishchuk/onlineschool/controller/dto/AuthResponse.java
package com.mishchuk.onlineschool.controller.dto;

import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Data;
import lombok.NoArgsConstructor;

import java.util.UUID;

@Data
@Builder
@AllArgsConstructor
@NoArgsConstructor
public class AuthResponse {
    private String accessToken;
    private UUID userId;
    private String role;
}