# src/main/resources/api.yaml
openapi: 3.0.3
info:
  title: Online School API
  description: REST API для платформи онлайн-навчання. Включає управління курсами, модулями, уроками, користувачами, реєстрацією на курси та файлами.
  version: 1.0.0
  contact:
    name: API Support
    email: support@onlineschool.com
servers:
  - url: http://localhost:8080
    description: Local development server
...
security:
  - bearerAuth: []
paths:
...
  /api/auth/register:
    post:
      tags:
        - Authentication
      summary: Реєстрація нового користувача
      operationId: register
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PersonCreateDto'
      responses:
        '200':
          description: Успішна реєстрація
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
          headers:
            Set-Cookie:
              description: Refresh token as HTTP-only cookie
              schema:
                type: string
                example: refreshToken=abc123; HttpOnly; Secure; Path=/; Max-Age=604800
        '400':
          description: Невалідні дані
  /api/auth/login:
    post:
      tags:
        - Authentication
      summary: Логін користувача
      operationId: login
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthRequest'
      responses:
        '200':
          description: Успішний логін
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
          headers:
            Set-Cookie:
              description: Refresh token as HTTP-only cookie
              schema:
                type: string
        '401':
          description: Невірні credentials
  /api/auth/refresh:
    post:
      tags:
        - Authentication
      summary: Оновлення access token
      operationId: refreshToken
      security: []
      parameters:
        - in: cookie
          name: refreshToken
          required: true
          schema:
            type: string
          description: Refresh token from cookie
      responses:
        '200':
          description: Новий access token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          description: Невалідний refresh token
  /api/auth/logout:
    post:
      tags:
        - Authentication
      summary: Вихід користувача
      operationId: logout
      responses:
        '200':
          description: Успішний logout
        '401':
          description: Unauthorized

  # ==================== Courses ====================
  /courses:
    get:
      tags:
        - Courses
      summary: Отримати всі курси
      operationId: getAllCourses
      security: []
      responses:
        '200':
          description: Список курсів
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CourseDto'
        '204':
          description: Немає курсів
    post:
      tags:
        - Courses
      summary: Створити новий курс
      operationId: createCourse
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CourseCreateDto'
      responses:
        '201':
          description: Курс створено
        '400':
          description: Невалідні дані
  /courses/{id}:
    get:
      tags:
        - Courses
      summary: Отримати курс за ID
      operationId: getCourse
      security: []
      parameters:
        - $ref: '#/components/parameters/IdPathParam'
      responses:
        '200':
          description: Деталі курсу
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CourseDto'
        '404':
          description: Курс не знайдено
    put:
      tags:
        - Courses
      summary: Оновити курс
      operationId: updateCourse
      parameters:
        - $ref: '#/components/parameters/IdPathParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CourseUpdateDto'
      responses:
        '204':
          description: Курс оновлено
        '404':
          description: Курс не знайдено
    delete:
      tags:
        - Courses
      summary: Видалити курс
      operationId: deleteCourse
      parameters:
        - $ref: '#/components/parameters/IdPathParam'
      responses:
        '204':
          description: Курс видалено

  # ==================== Modules ====================
...
  /persons:
    get:
      tags:
        - Persons
      summary: Отримати всіх користувачів
      operationId: getAllPersons
      responses:
        '200':
          description: Список користувачів
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PersonDto'
        '204':
          description: Немає користувачів
    post:
      tags:
        - Persons
      summary: Створити користувача
      operationId: createPerson
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PersonCreateDto'
      responses:
        '201':
          description: Користувача створено
  /persons/{id}:
    get:
      tags:
        - Persons
      summary: Отримати користувача за ID
      operationId: getPerson
      parameters:
        - $ref: '#/components/parameters/IdPathParam'
      responses:
        '200':
          description: Деталі користувача
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PersonDto'
        '404':
          description: Користувача не знайдено
    put:
      tags:
        - Persons
      summary: Оновити користувача
      operationId: updatePerson
      parameters:
        - $ref: '#/components/parameters/IdPathParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PersonUpdateDto'
      responses:
        '204':
          description: Користувача оновлено
        '404':
          description: Користувача не знайдено
    delete:
      tags:
        - Persons
      summary: Видалити користувача
      operationId: deletePerson
      parameters:
        - $ref: '#/components/parameters/IdPathParam'
      responses:
        '204':
          description: Користувача видалено



  # ==================== Files ====================
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    IdPathParam:
      in: path
      name: id
      required: true
      schema:
        type: string
        format: uuid
      description: UUID ідентифікатор
  schemas:
...
    AuthRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          example: user@example.com
        password:
          type: string
          format: password
          example: password123
    AuthResponse:
      type: object
      properties:
        accessToken:
          type: string
          description: JWT access token (15 хвилин)
        tokenType:
          type: string
          example: Bearer
        userId:
          type: string
          format: uuid
        email:
          type: string
          format: email
        role:
          type: string
          enum: [USER, ADMIN]

    # ==================== Person Schemas ====================
    PersonCreateDto:
      type: object
      required:
        - firstName
        - lastName
        - email
        - password
      properties:
        firstName:
          type: string
          example: John
        lastName:
          type: string
          example: Doe
        email:
          type: string
          format: email
          example: john.doe@example.com
        password:
          type: string
          format: password
          minLength: 6
    PersonDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
          format: email
        role:
          type: string
          enum: [USER, ADMIN]
        status:
          type: string
          enum: [ACTIVE, BLOCKED]
    PersonUpdateDto:
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
          format: email



    # ==================== Course Schemas ====================
    CourseCreateDto:
      type: object
      required:
        - name
        - description
      properties:
        name:
          type: string
          example: Introduction to Java
        description:
          type: string
          example: Learn Java programming from scratch
        status:
          type: string
          enum: [DRAFT, PUBLISHED, ARCHIVED]
          default: DRAFT
    CourseDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
        modulesNumber:
          type: integer
        status:
          type: string
          enum: [DRAFT, PUBLISHED, ARCHIVED]
    CourseUpdateDto:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        status:
          type: string
          enum: [DRAFT, PUBLISHED, ARCHIVED]

    # ==================== Module Schemas ====================
...